requirements.txt
Django==4.2.16
djangorestframework==3.16.0
djangorestframework-api-response==0.1.0
django-redis==5.4.0
django-environ==0.12.0
uvicorn==0.34.0
channels==4.1.0
channels-redis==4.2.0
whitenoise==6.9.0
psycopg2==2.9.4
pillow==11.0.0

django-storages==1.14.4
boto3==1.35.67
botocore==1.35.67

cryptography==41.0.7
rsa==4.9